+++ src/main/java/com/rocs/infirmary/application/controller/inventory/InventoryController.java
@@ -1,7 +1,7 @@
 package com.rocs.infirmary.application.controller.inventory;
 
 import com.rocs.infirmary.application.controller.helper.ControllerHelper;
-import com.rocs.infirmary.application.controller.lowstock.helper.NotificationAlertHelper;
+import com.rocs.infirmary.application.controller.notification.helper.NotificationAlertHelper;
 import com.rocs.infirmary.application.data.model.inventory.medicine.Medicine;
 import com.rocs.infirmary.application.module.inventory.management.application.InventoryManagementApplication;
 import javafx.collections.FXCollections;
@@ -252,7 +252,7 @@ public class InventoryController implements Initializable {
      * this method handles the refresh functionality for inventory table
      ***/
     public void refresh() {
-        NotificationAlertHelper.checkLowStockAndShowAlert();
+        NotificationAlertHelper.getNotification();
         selectAllCheckbox.setSelected(false);
         List<Medicine> medicineList = inventoryManagementApplication.getMedicineInventoryFacade().getAllMedicine();
         for (Medicine med : medicineList) {
