+++ src/main/java/module-info.java
@@ -14,7 +14,7 @@ module com.rocs.infirmaryapplication {
     opens com.rocs.infirmary.application.controller.dashboard to javafx.fxml;
     opens com.rocs.infirmary.application.data.model.person.student to javafx.fxml;
     opens com.rocs.infirmary.application.controller.mainpage to javafx.fxml;
-    opens com.rocs.infirmary.application.controller.lowstock to javafx.fxml;
+    opens com.rocs.infirmary.application.controller.notification to javafx.fxml;
 
     exports com.rocs.infirmary.application;
     exports com.rocs.infirmary.application.data.model.inventory;
@@ -34,7 +34,7 @@ module com.rocs.infirmaryapplication {
     opens com.rocs.infirmary.application.controller.records to javafx.fxml;
     exports com.rocs.infirmary.application.data.model.medicalrecord;
     opens com.rocs.infirmary.application.data.model.medicalrecord to javafx.fxml;
-    exports com.rocs.infirmary.application.controller.lowstock;
+    exports com.rocs.infirmary.application.controller.notification;
     opens com.rocs.infirmary.application.module.lowstock.notification.service.application;
     exports com.rocs.infirmary.application.module.lowstock.notification.service.application to javafx.fxml;
