+++ src/main/resources/styles/notification.css
@@ -6,7 +6,7 @@
     -fx-text-fill: #000000;
 }
 
-#alertCard {
+.alertCard {
     -fx-background-color: #FFE1CC;
     -fx-background-radius: 12px;
     -fx-padding: 25px;
@@ -16,13 +16,13 @@
       -fx-max-height: Infinity;
 }
 
-#alertTitle {
+.alertTitle {
     -fx-font-size: 14px;
     -fx-font-weight: bold;
     -fx-text-fill: #000000;
 }
 
-#alertMessage {
+.alertMessage {
     -fx-font-size: 12.5px;
     -fx-text-fill: #000000;
     -fx-line-spacing: 0.5em;
@@ -49,30 +49,6 @@
   -fx-min-height: 0;
   -fx-pref-height: -1;
   -fx-max-height: Infinity;
-}
-
-.alert-card {
-    -fx-background-color: #FFE1CC;
-        -fx-background-radius: 12px;
-        -fx-padding: 25px;
-
-        -fx-min-height: 0;
-          -fx-pref-height: -1;
-          -fx-max-height: Infinity;
-}
-
-.alert-title {
-    -fx-font-size: 14px;
-    -fx-font-weight: bold;
-    -fx-text-fill: #000000;
-}
-
-.alert-message {
-    -fx-font-size: 12.5px;
-    -fx-text-fill: #000000;
-    -fx-line-spacing: 0.5em;
-    -fx-wrap-text: true;
-    -fx-max-width: 600px;
 }
 
 .alert-close-button {
