+++ src/main/resources/views/LowStockNotificationModal.fxml
@@ -10,14 +10,14 @@
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
 
-<VBox fx:id="alertContainer" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="633.0" prefWidth="473.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.rocs.infirmary.application.controller.lowstock.NotificationController">
+<VBox fx:id="alertContainer" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="633.0" prefWidth="473.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.rocs.infirmary.application.controller.notification.NotificationController">
 
     <stylesheets>
         <URL value="@../styles/notification.css" />
         <URL value="@../styles/buttons.css" />
     </stylesheets>
    <children>
-      <HBox prefHeight="100.0" prefWidth="200.0">
+      <HBox prefHeight="50.0" prefWidth="200.0">
          <children>
             <Label fx:id="alertHeader" text="NOTIFICATION" />
             <Button fx:id="closeButton" mnemonicParsing="false" onMouseClicked="#onCloseButton" text="X" translateX="280.0" />
